First functional version

+++ Application/CommadLine.cs
@@ -16,13 +16,14 @@ public class CommandLineService : IInvokeService
             Console.WriteLine("Get a long URL (R):");
             Console.WriteLine("Get statistics of the short URL (S):");
             var response = Console.ReadLine();
-            switch (response.ToLower())
+            switch (response?.ToLower())
             {
                 case "a":
-                    AddUrl();   
+                    AddUrl().GetAwaiter().GetResult();   
                     break;
                 case "d":
-                    DeleteUrl();
+                    DeleteUrl().GetAwaiter()
+                        .GetResult();
                     break;
                 case "r":
                     GetUrl();
@@ -36,44 +37,87 @@ public class CommandLineService : IInvokeService
             }
 
             Console.WriteLine("Do you want to continue? (y/n)");
-        } while (Console.ReadLine().ToLower() == "y");
+        } while (Console.ReadLine()?.ToLower() == "y");
     }
 
-    private void AddUrl()
+    private async Task AddUrl()
     {
-        try 
+
+        Console.WriteLine("Enter the long URL:");
+        var longUrl = Console.ReadLine();
+        if (longUrl != null)
         {
-            Console.WriteLine("Enter the long URL:");
-            var longUrl = Console.ReadLine();
-            string shortUrl = _UrlSvc.ShortenUrl(longUrl);
-            Console.WriteLine($"Short URL: {shortUrl}");
+            try{
+                string shortUrl = await _UrlSvc.ShortenUrl(longUrl);
+                Console.WriteLine($"Short URL: {shortUrl}");
+            }   
+            catch (UriFormatException)
+            {
+                Console.WriteLine("Invalid URL");
+                return;
+            }
         }
-        catch (System.InvalidOperationException ex)
+        else
         {
-            Console.WriteLine(ex.Message);
+            Console.WriteLine("Invalid URL");
         }
     }
-    private void DeleteUrl()
+    private async Task DeleteUrl()
     {
         Console.WriteLine("Enter the short URL:");
         var shortUrlToDelete = Console.ReadLine();
-        _UrlSvc.DeleteUrl(shortUrlToDelete);
-        Console.WriteLine("URL deleted");
+        if(shortUrlToDelete == null)
+        {
+            Console.WriteLine("Invalid URL");
+            return;
+        }
+        try {
+            await _UrlSvc.DeleteUrl(shortUrlToDelete);
+        }
+        catch (System.InvalidOperationException)
+        {
+            Console.WriteLine("Short URL does not exist");
+            return;
+        }
     }   
-    private void GetUrl()
+    private async void GetUrl()
     {
         Console.WriteLine("Enter the short URL:");
         var shortUrlToGet = Console.ReadLine();
-        string longUrl = _UrlSvc.ExpandUrl(shortUrlToGet);
-        Console.WriteLine($"Long URL: {longUrl}");
+        if(shortUrlToGet == null)
+        {
+            Console.WriteLine("Invalid URL");
+            return;
+        }
+        try {
+            var longUrl = _UrlSvc.ExpandUrl(shortUrlToGet).GetAwaiter().GetResult();
+            Console.WriteLine($"Long URL: {longUrl}");
+        }
+        catch (System.InvalidOperationException)
+        {
+            Console.WriteLine("Short URL does not exist");
+            return;
+        }
     }
     private void GetStatistics()
     {
         Console.WriteLine("Enter the short URL:");
         var shortUrlToGetStats = Console.ReadLine();
-        var stats = _UrlSvc.GetStatistics(shortUrlToGetStats);
-        Console.WriteLine($"Short URL: {shortUrlToGetStats}");
-        Console.WriteLine($"Number of times accessed: {stats.NumberOfTimesAccessed}");
-        Console.WriteLine($"Last accessed on: {stats.LastAccessedOn}");
+        if(shortUrlToGetStats == null)
+        {
+            Console.WriteLine("Invalid URL");
+            return;
+        }
+        try {
+            var stats = _UrlSvc.GetStatistics(shortUrlToGetStats).GetAwaiter().GetResult();
+            Console.WriteLine($"Short URL: {shortUrlToGetStats}");
+            Console.WriteLine($"Number of times accessed: {stats.NumberOfTimesAccessed}");
+            Console.WriteLine($"Last accessed on: {stats.LastAccessedOn}");
+        }
+        catch (System.InvalidOperationException)
+        {
+            Console.WriteLine("Short URL does not exist");
+            return;
+        }
     }
 }
